bug: 38768
Mailbox IDs are now longs in ZCS.  The underlying MySQL database still uses an INTEGER datatype for the row, but the server can now theoretically handle a mailbox ID larger than MAXINT.

http://bugzilla.zimbra.com/show_bug.cgi?id=38768

Copied from Perforce
 Change: 161989

+++ src/java/com/zimbra/common/util/ZimbraLog.java
@@ -589,8 +589,8 @@ public class ZimbraLog {
     /**
      * Adds mailbox id to the current thread's logging context.
      */
-    public static void addMboxToContext(int mboxId) {
-        addToContext(C_MID, Integer.toString(mboxId));
+    public static void addMboxToContext(long mboxId) {
+        addToContext(C_MID, Long.toString(mboxId));
     }
     
     /**
